N2O-ConteinerGuard/backend/pyproject.toml
[tool.poetry]
name = "container-guard"
version = "0.1.0"
description = "N2O ContainerGuard backend service"
authors = ["ContainerGuard Team <security@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.0"
SQLAlchemy = "^2.0.29"
alembic = "^1.13.1"
httpx = "^0.27.0"
PyYAML = "^6.0.1"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.8.1"]
build-backend = "poetry.core.masonry.api"

